pyproject.toml
[tool.poetry]
name = "presidio-streamlit"
version = "0.1.0"
description = ""
authors = ["Presidio"]
readme = "index.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
presidio-analyzer = {extras = ["transformers"], version = "^2.2.355"}
presidio-anonymizer = "^2.2.355"
streamlit = "^1.39.0"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"
st-annotated-text = "^4.0.1"
streamlit-tags = "^1.2.8"
openai = "^1.53.0"
en_core_web_lg = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0.tar.gz"}
en_core_web_sm = {url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0.tar.gz"}

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

